requirements.txt
pika
robotframework